+++ tx-backend/src/test/java/org/eclipse/tractusx/traceability/integration/bpn/mapping/BpnMappingControllerIT.java
@@ -71,7 +71,7 @@ class BpnMappingControllerIT extends IntegrationTestSpecification {
     void givenBpnMappingRequest_whenUpdate_thenUpdateIt() throws JoseException {
         {
             // given
-            BpnMappingRequest mappings = new BpnMappingRequest("BPN123", "https://newurl.com");
+            BpnMappingRequest mappings = new BpnMappingRequest("BPNL00000003TEST", "https://newurl.com");
             bpnEdcMappingSupport.defaultBpnEdcMappingStored();
 
             // when
@@ -93,7 +93,7 @@ class BpnMappingControllerIT extends IntegrationTestSpecification {
                     .then()
                     .statusCode(200)
                     .body("", Matchers.hasSize(2))
-                    .body("[1].bpn", Matchers.equalTo("BPN123"))
+                    .body("[1].bpn", Matchers.equalTo("BPNL00000003TEST"))
                     .body("[1].url", Matchers.equalTo("https://newurl.com"));
         }
     }
@@ -108,7 +108,7 @@ class BpnMappingControllerIT extends IntegrationTestSpecification {
                 .contentType(ContentType.JSON)
                 .header(oAuth2Support.jwtAuthorization(ADMIN))
                 .when()
-                .delete("/api/bpn-config/BPN123")
+                .delete("/api/bpn-config/BPNL00000003TEST")
                 .then()
                 .statusCode(204);
 
@@ -121,8 +121,24 @@ class BpnMappingControllerIT extends IntegrationTestSpecification {
                 .then()
                 .statusCode(200)
                 .body("", Matchers.hasSize(1))
-                .body("[0].bpn", Matchers.equalTo("BPN456"))
+                .body("[0].bpn", Matchers.equalTo("BPNL00000002TEST"))
                 .body("[0].url", Matchers.equalTo("https://test456.de"));
+    }
+
+    @Test
+    void givenBpnMappingWrongBPNFormat_whenSave_thenBadRequest() throws JoseException {
+        // given
+        BpnMappingRequest mappings = new BpnMappingRequest("ABC", "http://localhost:12345/abc");
+
+        // when
+        given()
+                .contentType(ContentType.JSON)
+                .body(List.of(mappings))
+                .header(oAuth2Support.jwtAuthorization(ADMIN))
+                .when()
+                .post("/api/bpn-config")
+                .then()
+                .statusCode(400);
     }
 
     @Test
@@ -165,7 +181,7 @@ class BpnMappingControllerIT extends IntegrationTestSpecification {
     @Test
     void givenUserRole_whenCreateBpnMapping_thenReturn403() throws JoseException {
         // given
-        BpnMappingRequest request = new BpnMappingRequest("BPN123", "https://newurl.com");
+        BpnMappingRequest request = new BpnMappingRequest("BPNL00000003CSGF", "https://newurl.com");
 
         // when/then
         given()