+++ CHANGELOG.md
@@ -13,6 +13,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Sequence Diagrams for Data Provisioning Flow
 - Added User Manual for Data import
 - Added github action to connect pull request with github issue
+- Added bpn validator to bpn edc mapping
 
 ### Changed
 - Fixed security findings